move fetch helper to global level
