+++ views/menus/menus/github-menu.js
@@ -3,7 +3,7 @@ import React from 'react'
 import {View, Text} from 'react-native'
 import LoadingView from '../../components/loading'
 import {FilteredMenuView} from '../parts/filtered-menu'
-import {fetchJson} from './fetch'
+import {fetchJson} from '../../components/fetch'
 import type {TopLevelViewPropsType} from '../../types'
 import {TopLevelViewPropTypes} from '../../types'
 import type momentT from 'moment'
